Fix an infinite loop with cyclic constraints
There was a check to see if an edge existed *before* the generated
constraint was simplified and expanded to a list of edges.  The check
now occurs after the simplification.  Before the simplification step
eliminated some edges before they were inserted into the graph,
meaning the existence check always failed and generated a "new" edge.
This led to the infinite loop.