Parallelize an important part of the constraint graph saturation
The code finding new edges to add to the graph is now parallelized
using monad-par.  It has shown impressive speedups in an
implementation of Andersens analysis.

The tradeoff is that redundant edges may be generated compared to
before, but they are not added to the graph.

Currently, this relies on the Direct scheduler in monad-par 0.3.
There is a bug in the default scheduler that causes crashes in the
presence of nesting, which this code uses extensively.  Once the bug
is fixed, the version requirement should be bumped up and the code can
go back to the default scheduler.